requirements.txt
#Django packages
Django==1.11.6
djangorestframework==3.7.1